Flipped 'Socket' / 'WebSocket' definitions.
